Fix dtype error

+++ training.py
@@ -112,7 +112,7 @@ class DistillationLoss:
         self.alpha = alpha
 
     def __call__(self, y, labels, teacher_labels):
-        cross_entropy_hard = f.cross_entropy(y.double(), labels.double())
+        cross_entropy_hard = f.cross_entropy(y.double(), labels.long())
         cross_entropy_soft = nn.KLDivLoss()(f.log_softmax(y.double()/self.temperature, dim=-1),
                                             f.softmax(teacher_labels.double()/self.temperature, dim=-1))
         loss = self.alpha * cross_entropy_hard + (1 - self.alpha) * (self.temperature ** 2) * cross_entropy_soft
